+++ src/main/java/de/danielbasedow/prospecter/core/QueryPosting.java
@@ -4,15 +4,15 @@ package de.danielbasedow.prospecter.core;
  * Immutable representation of query posting (queryId, bitToToggle)
  */
 public class QueryPosting {
-    private long queryId;
+    private int queryId;
     private short queryBit;
 
-    public QueryPosting(long queryId, short queryBit) {
+    public QueryPosting(int queryId, short queryBit) {
         this.queryId = queryId;
         this.queryBit = queryBit;
     }
 
-    public long getQueryId() {
+    public int getQueryId() {
         return queryId;
     }
